py-client/requirements.txt
pyyaml
flask
waitress
flask_cors
itsdangerous
validators
#pyinstaller==4.9